# Conflicts:
#	gradlew
#	gradlew.bat
#	src/main/java/hanghae99/clonecoding/airbnb/controller/HotelController.java
#	src/main/java/hanghae99/clonecoding/airbnb/service/HotelService.java

+++ src/main/java/hanghae99/clonecoding/airbnb/security/JwtProvider.java
@@ -51,7 +51,7 @@ public class JwtProvider {
         return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
     }
     public MemberDetail getMemberDetail(String token){
-        long id = Long.parseLong(getUserPk(token));
+        int id = Integer.parseInt(getUserPk(token));
         Member member = repo.findById(id).get();
         return new MemberDetail(member);
     }
